tools: score_chroma: ignore bass PCP
Do not account bass half of PCP/chromagram for now, treat all chords as if
in root position. Later, when treble PCPs are clear enough, bass PCP shall be
considered again.

The reason is that bass PCP of real chord is never similar to the one from
template, so PSNR calculation become too biased and prevent finer tuning of
PCP generating process. Also, templates provide inversions for major/minor
chords only, because of that great deal of 2374-set was unused.